Rename package name from ssohandler to sso
