+++ pkg/auth/handler/sso/auth_handler.go
@@ -1,4 +1,4 @@
-package ssohandler
+package sso
 
 import (
 	"encoding/base64"
