+++ pkg/auth/handler/sso/auth_url.go
@@ -1,4 +1,4 @@
-package ssohandler
+package sso
 
 import (
 	"encoding/json"
