+++ pkg/auth/handler/sso/auth_url_test.go
@@ -1,4 +1,4 @@
-package ssohandler
+package sso
 
 import (
 	"net/url"
