+++ pkg/auth/handler/sso/config_test.go
@@ -1,4 +1,4 @@
-package ssohandler
+package sso
 
 import (
 	"net/http"
