+++ pkg/auth/handler/sso/customtoken.go
@@ -1,4 +1,4 @@
-package ssohandler
+package sso
 
 import (
 	"encoding/json"
