+++ pkg/auth/handler/sso/customtoken_test.go
@@ -1,4 +1,4 @@
-package ssohandler
+package sso
 
 import (
 	"fmt"
