+++ pkg/auth/handler/sso/handler.go
@@ -1,4 +1,4 @@
-package ssohandler
+package sso
 
 import (
 	"time"
