+++ pkg/auth/handler/sso/iframe_handler.go
@@ -1,4 +1,4 @@
-package ssohandler
+package sso
 
 import (
 	"io"
