+++ pkg/auth/handler/sso/iframe_handler_test.go
@@ -1,4 +1,4 @@
-package ssohandler
+package sso
 
 import (
 	"net/http"
